refactor Header.jsx & Styles.jsx
