+++ src/components/Signup.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
-import Header from '../components/Header.jsx'
+import Header from '../components/Header'
 import Footer from './Footer.jsx'
 import { Link, useNavigate } from 'react-router-dom'
-import { StyledContainer, SignupDiv, TittleText, FormContainerSignup, SignupDivButton, InlineText, GridSignup } from '../Styles.jsx'
+import { StyledContainer, SignupDiv, TittleText, FormContainerSignup, SignupDivButton, InlineText, GridSignup } from './styles'
 import { Text, Grid, FormControl, FormLabel, InputLeftElement, InputRightElement, Input, InputGroup, Button, useToast, GridItem } from '@chakra-ui/react'
 import { BsPencilSquare, BsPerson } from 'react-icons/bs'
 import { HiOutlineIdentification } from 'react-icons/hi'
